+++ libs/web/notification/data-access/src/lib/models/notification.model.ts
@@ -1,0 +1,8 @@
+import { Nullable } from '@teaching-scheduling-system/core/data-access/models';
+import { EchoMessage } from './echo.model';
+
+export interface NotificationPage {
+  data: EchoMessage[];
+  hasNext: boolean;
+  milestone: Nullable<string>;
+}